refactor: тип переменных для обозначения курса валют заменен с double на BigDecimal

+++ src/main/java/com/example/exchange_rate/model/ExchangeRate.java
@@ -2,6 +2,7 @@ package com.example.exchange_rate.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Map;
 
@@ -9,5 +10,5 @@ import java.util.Map;
 public class ExchangeRate {
     private String base;
     private LocalDate date;
-    private Map<String, Double> rates;
+    private Map<String, BigDecimal> rates;
 }
